document + refactor numcomplex/bigint dependencies
