+++ Cargo.toml
@@ -38,6 +38,8 @@ proptest = { version = "0.10.1", default-features = false, features = ["std"] }
 # features needed to run the PyO3 test suite
 pyo3 = { path = ".", default-features = false, features = ["macros", "auto-initialize"] }
 serde_json = "1.0.61"
+# needed for num-complex doctest
+nalgebra = "0.27.1"
 
 [build-dependencies]
 pyo3-build-config = { path = "pyo3-build-config", version = "=0.14.0-alpha.0" }
