+++ src/num_bigint.rs
@@ -1,3 +1,7 @@
+// Copyright (c) 2017-present PyO3 Project and Contributors
+//
+// based on Daniel Grunwald's https://github.com/dgrunwald/rust-cpython
+
 #![cfg(all(feature = "num-bigint", not(any(Py_LIMITED_API, PyPy))))]
 #![cfg_attr(
     docsrs,
@@ -5,8 +9,7 @@
 )]
 //!  Conversions to and from [num-bigint](https://docs.rs/num-bigint)’s [`BigInt`] and [`BigUint`] types.
 //!
-//! This is useful for converting Python integers, which have arbitrary precision,
-//! when they may not fit in Rust's built-in integer types.
+//! This is useful for converting Python integers when they may not fit in Rust's built-in integer types.
 //!
 //! # Setup
 //!
@@ -15,7 +18,7 @@
 //! ```toml
 //! [dependencies]
 //! num-bigint = "0.4"
-//! pyo3 = { version = "0.14.0", features = ["num-bigint"] }
+//! pyo3 = { version = "0.14", features = ["num-bigint"] }
 //! ```
 //!
 //! Note that you must use compatible versions of num-bigint and PyO3.
@@ -23,18 +26,20 @@
 //!
 //! ## Examples
 //!
-//! [`BigInt`] and [`BigUint`] can be used represent arbitrary precision integers:
+//! Using [`BigInt`] to correctly increment an arbitrary precision integer.
+//! This is not possible with Rust's native integers if the Python integer is too large,
+//! in which case it will fail its conversion and raise `OverflowError`.
 //!
 //! ```rust
 //! use num_bigint::BigInt;
 //! use pyo3::prelude::*;
 //! use pyo3::wrap_pyfunction;
-//! 
+//!
 //! #[pyfunction]
 //! fn add_one(n: BigInt) -> BigInt {
 //!     n + 1
 //! }
-//! 
+//!
 //! #[pymodule]
 //! fn my_module(_py: Python, m: &PyModule) -> PyResult<()> {
 //!     m.add_function(wrap_pyfunction!(add_one, m)?)?;
@@ -45,13 +50,12 @@
 //! Python code:
 //! ```python
 //! from my_module import add_one
-//! 
+//!
 //! n = 1 << 1337
 //! value = add_one(n)
-//! 
+//!
 //! assert n + 1 == value
 //! ```
-
 
 use crate::{
     err, ffi, types::*, AsPyPointer, FromPyObject, IntoPy, Py, PyAny, PyErr, PyNativeType,
